Move common code into middleware utils to make creation of new middleware simpler
